Suppress initgroups reqs for non-TACACS+ users
Generally the nsswitch.conf configuration will result in the sss NSS
service being queried after local sources, eg. see
debian/libnss-sss.postinst.

NSS supports the "initgroups" database which is used to obtain the
supplementary group list for a user. When there is no "initgroups"
service configuration in nsswitch.conf NSS will query all services
listed for the "groups" database - it will not return after a single
successful query, so that supplementary groups may be supplied by
multiple services.

By default, on Debian at least, there is no explicit "initgroups"
configuration. The result of this is that an "initgroups" request,
eg. by a call to getgrouplist(), will always query the compat
(ie. local) and sss services. Therefore TACACS+ session authorization
requests are issued for local users which are unknown to the server.

We do not support additional group memberships to be supplied by the
TACACS+ server so there is no need for these authorization requests.
Therefore prior to issuing one as a result of an "initgroups" NSS
request we check whether the user already exists in another database,
by a call to getpwnam(); if so we return without a TACACS+
transaction.

VRVDR-49750