tacplus: Check UID in range for initgroups req
The previous change to suppress initgroups requests for non-TACACS+
users in 8160c1bf34adcc0423c4b4ef9757042bd220795a also caused the
requests to be suppressed for TACACS+ users if they were present in
the SSS NSS service in-memory cache, since getpwnam() would return
an entry for those users in this case.

Therefore we now check the UID of any found user is outside the UID
range for the tacplus domain. The initgroups request is only
suppressed if the user is outside this range.

VRVDR-50035