tacplus: conn check on local user initgroups req
The previous change to suppress TACACS+ authorization requests in
response to an initgroups request prevented local fallback user
login, when TACACS+ is unavailable, from being permitted simply by
attempting login as that local user. This happened since the failure
of the now suppressed authorization request was what triggered the
tacplus provider to go offline and permit local fallback user logins.

To address this we now perform a connection check (ie. can we open a
connection to a TACACS+ server) in response to an initgroups request
for a local user. If the check fails then the tacplus provider is
taken offline. Thus local fallback logins can be permitted without
performing an actual TACACS+ transaction.

For an initgroups request for a system user (UID < 1000) we still
suppress all TACACS+ operations since they are not required.

VRVDR-50035