Call data provider on pam_close_session()
The PAM responder's implementation of pam_sm_close_session() now
calls the provider's session handler so the provider can know when a
session is terminated. Previously this handler was only called by
pam_sm_open_session().

In the case of the tacplus provider this means we can issue a TACACS+
accounting stop record.