Propagate provider error to PAM
Currently if an identity provider experiences an error on an
initgroups request from the PAM responder, the status returned to the
PAM stack will be PAM_USER_UNKNOWN (by a cache lookup returning no
result in cache_req_search_done(), thus returning ENOENT to
pam_dp_send_acct_req_done()).

In order to support configuring the system to allow local logins only
when TACACS+ is unavailable we need to return PAM_SYSTEM_ERR in this
situation.

This is achieved, in patch cache-req-propagate-dp-error.patch, by the
addition of an option on the cache_req_data struct which indicates
that the requester of a given search wishes to have any provider error
propagated back to them.

In propagate-initgroup-dp-failure-to-pam.patch we set this option for
INITGROUPS requests from the PAM responder and update
pam_dp_send_acct_req_done() to handle return values other than EOK
and ENOENT.

In propagate-provider-error.patch a new domain configuration option
is provided which enables the propagation of errors from the
providers of that domain. There is therefore no behaviour change by
default since to have provider errors propagated back to the PAM
responder this option must be enabled on at least one domain.

The set of patches in this commit provide a replacement for
propagate_pam_error_via_initgroup_failure.patch from the 1.12.x
branch.