Revert removal of local-user-name authz attr
Support for the "local-user-name=" TACACS+ authorization attribute
was previously removed in 19fbc3646489f1944f731cf2f2d740512286c5e1.

This commit adds that same support back so that the attribute remains
supported for releases using the 1.12.x branch. It will be removed in
a later rebase to a new version of SSSD.