tacplus: allocate persistent UIDs
Currently it is possible for a TACACS+ user to be allocated multiple
UIDs if they experience a session authorization failure.

When this occurs the user is immediately deleted from the sysdb.
If the user is subsequently permitted access to the system again
they will receive a new generated UID on login. This likely means
they will be unable to write to their home directory.

To avoid this we store an additional entry in the sysdb when we
first create a user containing their name and the UID they were
allocated. This entry is located under the tacplus_uid_allocation
subtree and remains even if the corresponding user entry is removed
from the sysdb. When storing a user in the sysdb we lookup whether
we have already allocated them a UID, ensuring that it remains
consistent for the lifetime of the sysdb.

We now also use the allocated UID to perform session authorization
requests in the getpwuid() handler. In the case where the UID being
looked up has been previously allocated, and the user it was
allocated to does not exist in the sysdb, we call get_pw_name() to
do a session authorization and add the user back to the sysdb if it
succeeds. This allows recovery of a user's existing sessions if they
remained open following the previous authorization failure (ie. we
will recover gracefully from a transient authorization failure).
Previously the use of a new UID in this case would result in existing
sessions remaining broken.

As before UID allocations will not persist a reboot.

VRVDR-49990