Make tacplus_can_connect() non-blocking
This change allows the tacplus ID provider to quickly return to a
responder that it is offline, while it is performing a connection
check via tacplus_can_connect(). Previously the provider would be
blocked on the result of the connection check and therefore could
not respond to requests from responders.

In cases where the provider is offline and the connection check is
also failing this removes significant delays in identity lookup
operations, which can be observed by users eg. as delays in spawning
processes.

In future all interaction with tacplusd over DBus should be made
non-blocking. For now, converting tacplus_can_connect() is a good
start.