Log local-user-name deprecation at warning level
When building on Debian 10 (implying a release of 2005 or later)
set the level on the syslog message warning of the deprecation and
future removal of support for the local-user-name authorization
attribute to warning.

On earlier releases the level remains at notice.

Support for local-user-name has been removed from the danos/1.16.x
branch.

VRVDR-50681