Move DBs, memcache, and pipes to /run
On DANOS we do not expect/require databases to persist a reboot. In
fact we remove any old databases on boot. Therefore we are free to
move them to a tmpfs.

This gives us better resiliency against a system failure where the
filesystem becomes readonly. It should also improve performance and
reduce disk wear, although likely not to a noticeable degree.

VRVDR-52319