monitor: Add "switch_user" configuration
Typically the sssd monitor process runs as root, allowing it to start
child processes with root permissions. However in some configurations
it is not required to have these elevated permissions.

In these scenarios the switch_user boolean configuration can be used
to instruct the monitor process to switch to the user specified in
the configuration during startup, via server_setup(). This switch
occurs before child processes are started and the children are
therefore started as this user.

This option might break some SSSD configurations which require the
child domain/service processes to be started as root. However no
issues have been observed with the configurations supported by DANOS.

However one known gotcha when setting switch_user to True and running
SSSD as a non-root user is that the pam service must be socket
activated, and not set in the "services" configuration. This ensures
that the service's sockets are owned by root, which is required, even
though the process will be executed as an un-privileged user.

An ExecPreStart statement is also added to sssd.service.in, similar
to those specified in the per-component services, to ensure correct
ownership of all SSSD log files.

VRVDR-52730