IFP: Allow responder to run as non-root user
Update the sssd-ifp service template to handle @SSSD_USER@
substitution, causing sssd-ifp to run as the same user as other
SSSD components.

This also requires the DBus policy to be adjusted to allow the
specified user to own the service name on the system bus. Therefore
a template of the policy is created, allowing @SSSD_USER@
substitution.

The --uid and --gid arguments were accepted by sssd_ifp but they had
no effect and are therefore removed and no longer specified in the
systemd/DBus services.

server_setup() is also updated so that it does not attempt to switch
users for a DBus activated service (mirroring the behaviour for
socket activated services). Otherwise sssd_ifp will attempt to switch
to the root user at startup and fail.

Finally, mirroring the other sssd component services, we use
ExecStartPre to ensure correct ownership of the component log file.

VRVDR-52730