tacplus: Use pam_data->conn_info for rem_addr
We always wish to use an IP address in preference to a hostname in
the rem_addr field of a TACACS+ accounting request. Previously we
used pam_data->rhost which is populated from PAM_RHOST, as set by the
PAM application. Depending upon the DNS environment the value of
PAM_RHOST can be a hostname.

Therefore we now use the source address from pam_data->conn_info, if
it is available. Otherwise we fallback to using pam_data->rhost.

VRVDR-53114