whitespace_test: fix exclusion of debian directory
These tests are supposed to ignore trailing whitespace in files in
the debian directory. However this is not the case with the current
exclude pattern and usage of `git grep`.

Use --full-name with `git grep` to ensure its output is always
relative to the repository root, rather than the current directory.
Additionally remove the leading slash from the exclude pattern as
this will never match the output.

VRVDR-53874